+++ src/frontend/js/utils/errors/handle.spec.ts
@@ -2,25 +2,25 @@
  * @jest-environment jsdom
  */
 import * as Sentry from '@sentry/browser';
-import { ContextFactory } from 'utils/test/factories';
+import { ContextFactory as mockContextFactory } from 'utils/test/factories';
+import { handle } from './handle';
 
+jest.mock('utils/context', () => ({
+  __esModule: true,
+  default: mockContextFactory({
+    sentry_dsn: 'https://sentry.local.test',
+  }).generate(),
+}));
 jest.mock('@sentry/browser');
 
 describe('handle', () => {
-  window.__richie_frontend_context__ = {
-    context: ContextFactory({
-      sentry_dsn: 'https://sentry.local.test',
-    }).generate(),
-  };
-  const { handle } = require('./handle');
-
   it('should initialize sentry', () => {
     expect(Sentry.init).toBeCalledTimes(1);
     expect(Sentry.configureScope).toBeCalledTimes(1);
   });
 
   it('should report error to sentry', () => {
-    handle('An error for test');
+    handle(new Error('An error for test'));
     expect(Sentry.captureException).toBeCalledTimes(1);
   });
 });